Upd_120421 * Class DumpTable updated - updated class structure, more fields discovered, implemented basic methods. * Class Config updated - added 'GetcollmatMaterialId' function. * Class BuiltinType updated - added implementation for 'GetFacecollMaterialId', 'GetMessageId' and dump table related functions.

+++ TOD1/BuiltinType.h
@@ -134,7 +134,7 @@ public:
 	void	DumptableAddStringColumn(int*);	//	@485920
 	void	DumptableSetNumRows(int*);	//	@485950
 	void	DumptableSetIntegerValue(int*);	//	@485970
-	void	DumptableSetNumberValue(float*);	//	@4859A0
+	void	DumptableSetNumberValue(int*);	//	@4859A0
 	void	DumptableSetStringValue(int*);	//	@4859D0
 	void	DumptableWriteToFile(int*);	//	@485A00
 	void	DumptableClose(int*);	//	@485A90
